requirements.txt
bs4
cchardet
google_images_download
img2pdf
lxml
matplotlib
natsort
numpy
pillow
spacy
tensorflow-gpu>=1.5.0
cython
https://api.github.com/repos/thtrieu/darkflow/tarball
https://api.github.com/repos/mwilliammyers/tensornets/tarball
